+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Ethernet Network
 - FC Network
 - FCOE Network
+- Hypervisor Cluster Profile
 - Hypervisor Manager
 - Interconnect
 - Interconnect Type